+++ pkg/cluster/helpers.go
@@ -36,7 +36,7 @@ func annVars(metadata v1.ObjectMeta) []bitesize.Annotation {
 	for k, v := range annotations {
 		var a bitesize.Annotation
 		a = bitesize.Annotation{
-			Name: annotations[k],
+			Name:  annotations[k],
 			Value: annotations[v],
 		}
 		retval = append(retval, a)